updatedStratClass

+++ src/main/java/StratPackage/Stratclass.java
@@ -20,7 +20,7 @@ public class Stratclass {
     // opening browser
     @BeforeMethod
     public void browseropen() {
-       // if (System.getProperty("browser").equalsIgnoreCase("chrome")) {
+       if (System.getProperty("browser").equalsIgnoreCase("chrome")) {
             WebDriverManager.chromedriver().setup();
             ChromeOptions options = new ChromeOptions();
             options.setPageLoadStrategy(PageLoadStrategy.NONE);
@@ -31,29 +31,28 @@ public class Stratclass {
             options.addArguments("--disable-browser-side-navigation");
             options.addArguments("--disable-gpu");
             driver = new ChromeDriver(options);
-//        } else if (System.getProperty("browser").equalsIgnoreCase("firefox")) {
-//            WebDriverManager.firefoxdriver().setup();
-//            FirefoxOptions options = new FirefoxOptions();
-//            options.setPageLoadStrategy(PageLoadStrategy.NONE);
-//            options.addArguments("enable-automation");
-//            options.addArguments("--no-sandbox");
-//            options.addArguments("--disable-infobars");
-//            options.addArguments("--disable-dev-shm-usage");
-//            options.addArguments("--disable-browser-side-navigation");
-//            options.addArguments("--disable-gpu");
-//            driver = new FirefoxDriver(options);
-//        } else if (System.getProperty("browser").equalsIgnoreCase("IE")) {
-//            WebDriverManager.iedriver().setup();
-//            InternetExplorerOptions options = new InternetExplorerOptions();
-//            options.setPageLoadStrategy(PageLoadStrategy.NONE);
+        } else if (System.getProperty("browser").equalsIgnoreCase("firefox")) {
+            WebDriverManager.firefoxdriver().setup();
+            FirefoxOptions options = new FirefoxOptions();
+            options.setPageLoadStrategy(PageLoadStrategy.NONE);
+            options.addArguments("enable-automation");
+            options.addArguments("--no-sandbox");
+            options.addArguments("--disable-infobars");
+            options.addArguments("--disable-dev-shm-usage");
+            options.addArguments("--disable-browser-side-navigation");
+            options.addArguments("--disable-gpu");
+            driver = new FirefoxDriver(options);
+        } else if (System.getProperty("browser").equalsIgnoreCase("IE")) {
+            WebDriverManager.iedriver().setup();
+            InternetExplorerOptions options = new InternetExplorerOptions();
+            options.setPageLoadStrategy(PageLoadStrategy.NONE);
 ////                                  options.addArguments("enable-automation");
 //                                  options.addArguments("--no-sandbox");
 //                                   options.addArguments("--disable-infobars");
 //                                   options.addArguments("--disable-dev-shm-usage");
 //                                   options.addArguments("--disable-browser-side-navigation");
-            //                                 options.addArguments("--disable-gpu");
-            //   driver = new InternetExplorerDriver(options);
-            // }
+   //                                         options.addArguments("--disable-gpu");driver = new InternetExplorerDriver(options);
+             }
             driver.get("https://www.hotukdeals.com/?page=4");
             driver.manage().window().maximize();
             driver.manage().deleteAllCookies();
